proxy/mirror/requirements.txt
huggingface_hub>=0.24.0
minio>=7.2.7
boto3>=1.34.0
python-dotenv>=1.0.1
tqdm>=4.66.0
requests>=2.31.0
